% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotationList.R
\docType{class}
\name{AnnotationList-class}
\alias{AnnotationList}
\alias{AnnotationList-class}
\title{The annotation data.table should have columns "accession" and "category"}
\description{
The annotation data.table should have columns "accession" and "category"
}

