% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotationList.R
\name{annotationMatrix}
\alias{annotationMatrix}
\title{Annotation Matrix}
\usage{
annotationMatrix(object)
}
\arguments{
\item{object}{An object which inherits from AnnotationList}
}
\value{
matrix A matrix of n=number of proteins rows and m=number of categories
columns
}
\description{
\code{annotationMatrix} produces the Cij matrix, where position i,j is 1 if gene i
can be annotated with category j.
}

