% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotationList.R
\name{annotationWeights}
\alias{annotationWeights}
\title{Annotation weights}
\usage{
annotationWeights(object)
}
\arguments{
\item{object}{An object which inherits from AnnotationList}
}
\value{
The data.table Dj holding the annotation weights.
}
\description{
\code{annotationWeights} provides the weights for each annotation
category. In the case of GO, this could be the distance from
the root object. Given this, all weights should be integers > 0.
In terms of the LP problem, the returned vector is equivalent to Dj.
The implementation needs to make sure that the order returned is consistent
with other outputs.
}

