% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gurobi.R
\name{obtainAnnotatedSolution}
\alias{obtainAnnotatedSolution}
\title{Obtain annotated solution}
\usage{
obtainAnnotatedSolution(gurobiRes, annotationList)
}
\arguments{
\item{gurobiRes}{The result of invoking gurobi() with the model produced by \code{produceGurobiModel}.}

\item{annotationList}{The same AnnotationList object provided to \code{produceGurobiModel}.}
}
\value{
a list containing \code{resMat}, which contains the result matrix (Terms as rows, Proteins as Columns,
a 1 in i,j means Term i is annotation Protein j), and \code{coverage}, which contains a summary per Term of
the maximum depth of the term and the number of proteins it is annotating (the coverage).
}
\description{
Produces an annotated version of the optimal within the gurobi result given (the result of calling gurobi()).
The annotation is extracted from the annotation list object.
}

