% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gurobi.R
\name{produceGurobiModel}
\alias{produceGurobiModel}
\title{Produce Gurobi Model}
\usage{
produceGurobiModel(annotationList, maxAnnotationTerms = 20,
  minDistanceFromRoot = 3)
}
\arguments{
\item{annotationList}{An AnnotationList object, from where the annotation matrix and
dimensions are obtained.}

\item{maxAnnotationTerms}{An integer with the maximum numbers of terms that should be used
for the overall annotation. Defaults to 20.}

\item{minDistanceFromRoot}{An integer with the minimum distance that terms need to have from
the root of the ontology to be eligible. Defaults to 3.}
}
\value{
the gurobi model, ready to be obtimized using gurobi(). It is a list containing
the A matrix, the objective function, the b vector, the senses vector, the right hand side,
the nature of the variables and the type (maximization.)
}
\description{
Produces a formatted gurobi model based on the annotation list object given and
the additional parameters that control the problem definition.
}

