% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AnnotationList.R
\name{retrieveAnnotation}
\alias{retrieveAnnotation}
\title{Retrieve Annotation}
\usage{
retrieveAnnotation(object)
}
\arguments{
\item{object}{An object which inherits from AnnotationList}
}
\value{
A new object of the same class with the annotation filled.
}
\description{
\code{retrieveAnnotation} Uses a data source (depending on
implementation) to retrieve annotations for the
defined set of proteins. The result is stored internally.
}
\details{
For the case of deep DAGs like the gene ontology, this method
is expected to provide the entire list of GOs from root to tip
that can be assigned to the proteins.
}

